Make progress in documentation and bookmarks
